ImGui interface for eye and look center

+++ GPA_2024_final_framework_indoor/Rendering_Framework/src/SceneManager.h
@@ -56,6 +56,6 @@ public:
 	int m_fs_diffuseAlbedo;
 	int m_fs_useTexture;
 
-	int m_test_meshcount = 0;
+	float m_test_meshcount = 0;
 };
